requirements.txt
pymongo
pandas

python-environ